requirements.txt
beautifulsoup4==4.13.4
langchain==0.3.25
langchain_core==0.3.59
langchain_openai==0.3.16
langchain_community
langgraph==0.4.3
pandas==2.2.3
pydantic==2.11.4
python-dotenv==1.1.0
python_docx==1.1.2
openpyxl
Requests==2.32.3
streamlit==1.44.1
google-search-results
